fix services and view

+++ color_extractor/views.py
@@ -99,11 +99,15 @@ def images(request):
 
 
 def delete_image(request, image_id):
-    try:
-        image = ImageExtraction.objects.get(id=image_id)
-        previous_image = ImageExtraction.objects.filter(id__lt=image_id).last()
-        image.delete()
-        previous_image.delete()
-    except ImageExtraction.DoesNotExist:
-        pass
+    s = request.session.get("user")
+    if s:
+        try:
+            image = ImageExtraction.objects.get(id=image_id)
+            previous_image = ImageExtraction.objects.filter(
+                id__lt=image_id, image_name=image.image_name
+            ).last()
+            image.delete()
+            previous_image.delete()
+        except ImageExtraction.DoesNotExist:
+            pass
     return redirect("images")
